Crate hello_world

Structs

LoanAnalyticsContract

LoanAnalyticsContractArgs

LoanAnalyticsContractArgs is a type for building arg lists for functions defined in “LoanAnalyticsContract”.

LoanAnalyticsContractClient

LoanAnalyticsContractClient is a client for calling the contract defined in “LoanAnalyticsContract”.

LoanStats

Statics

__SPEC_XDR_TYPE_LOANSTATS