hello_world

Function run

pub(crate) fn run(terminal: &mut DefaultTerminal) -> Result<()>

Run the application loop. This is where you would handle events and update the application state. This example exits when the user presses ‘q’. Other styles of application loops are possible, for example, you could have multiple application states and switch between them based on events, or you could have a single application state and update it based on events.